+++ c_impl/heap.h
@@ -9,7 +9,7 @@ typedef struct {
     size_t item_size;
     size_t count;
     size_t capacity;
-    int (*score)(const void*);
+    double (*score)(const void*);
 } Heap;
 
 extern char* _heap_popped_bytes;
@@ -59,7 +59,7 @@ void heap_build_impl(Heap* heap);
 void heap_push_impl(Heap* heap, char* bytes);
 void heap_append_impl(Heap* heap, char* bytes);
 char* heap_pop_impl(Heap* heap);
-void heap_init(Heap* heap, char* items_bytes, size_t count, size_t capacity, int (*score) (const void*));
+void heap_init(Heap* heap, char* items_bytes, size_t count, size_t capacity, double (*score) (const void*));
 void heap_free(Heap* heap);
 
 #endif
